requirements.txt
pillow
pytesseract
pandas
numpy
tqdm